Educhatbot/requirements.txt
python-multipart
uvicorn==0.25.0
urllib3
requests
mangum==0.17.0
langchain==0.1.0
PyPDF2
PyMuPDF==1.23.9
pylatexenc==2.10
boto3==1.34.19
azure-cognitiveservices-speech==1.34.1
langchain-community==0.0.13
openai==1.30.1
pydantic==2.6.4
fastapi==0.109.0
# pyautogen==0.2.26
pyautogen
tiktoken==0.5.2
opensearch-py==2.4.2
pandas==2.2.0
elevenlabs==0.2.27
pymysql==1.1.0
nltk==3.8.1
xlsxwriter==3.2.0
pydub==0.25.1
inflect==7.0.0
yt-dlp
did-api
bs4
unstructured